progressbar filters moved to new file in utils
