+++ trojsten/polls/templates/trojsten/polls/parts/progress.html
@@ -1,4 +1,4 @@
-{% load polls_parts staticfiles i18n %}
+{% load polls_parts staticfiles i18n progressbar %}
 {% if remaining.days >= 0 %}
